+++ src/locales/ko/common.json
@@ -235,5 +235,46 @@
       "ai": "AI/ML",
       "optimization": "최적화"
     }
+  },
+  "contact": {
+    "toast": {
+      "success": "성공적으로 전송되었습니다.",
+      "error": "전송에 실패했습니다. 다시 시도해주세요."
+    },
+    "form": {
+      "name": {
+        "label": "이름",
+        "placeholder": "이름을 입력해주세요",
+        "required": "이름을 입력해주세요."
+      },
+      "company": {
+        "label": "회사",
+        "placeholder": "회사명을 입력해주세요 (선택)"
+      },
+      "email": {
+        "label": "이메일",
+        "placeholder": "이메일을 입력해주세요",
+        "required": "이메일을 입력해주세요."
+      },
+      "title": {
+        "label": "제목",
+        "placeholder": "제목을 입력해주세요",
+        "required": "제목을 입력해주세요."
+      },
+      "message": {
+        "label": "메시지",
+        "placeholder": "메시지를 입력해주세요",
+        "required": "메시지를 입력해주세요."
+      },
+      "submit": "전송",
+      "sending": "전송 중..."
+    },
+    "info": {
+      "kakao": "카카오톡",
+      "openChat": "오픈채팅",
+      "email": "이메일",
+      "location": "위치",
+      "locationValue": "서울, 대한민국"
+    }
   }
 }
